requirements.txt
asgiref==3.11.0
attrs==25.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap5==2024.10
cryptography==44.0.1
defusedxml==0.7.1
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==6.0.1
django-allauth==65.14.0
django-auditlog==3.0.0
django-cache-url==3.4.5
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-log-viewer==1.1.8
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
environs==14.1.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
marshmallow==3.26.2
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
pillow==11.3.0
pip-review==1.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.22.3
shapely==2.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.6.3
uvicorn==0.34.0
whitenoise==6.8.2